placed css in separate folder in src
