+++ front-emeM/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import { BrowserRouter as Router } from 'react-router-dom'
 import 'semantic-ui-css/semantic.min.css'
-import './index.css';
+import './css/index.css';
 
 ReactDOM.render((
   <Router>
